Upgrades upgrades upgrades

+++ scripts/python-tools.sh
@@ -37,10 +37,28 @@ echo "Creating new virtualenv and upgrading pip"
 
 echo "Installing the Python tools"
 
-~/.ve/pytools/bin/pip install -U flake8 black flake8-bugbear pyflakes pylint isort
+~/.ve/pytools/bin/pip install -U \
+    flake8 \
+    black \
+    flake8-bugbear \
+    pyflakes \
+    pylint \
+    isort \
+    mypy
 
-for bin in flake8 black blackd pyflakes pylint isort;
+~/.ve/pytools/bin/pip install -U flake8 black flake8-bugbear pyflakes pylint mypy
+
+for bin in \
+    flake8 \
+    black \
+    blackd \
+    pyflakes \
+    pylint \
+    isort \
+    mypy;
 do
     echo "Linking $bin into ~/Applications/"
-    ln -s ~/.ve/pytools/bin/$bin ~/Applications/$bin
+    if [ ! -e ~/Applications/$bin ]; then
+        ln -s ~/.ve/pytools/bin/$bin ~/Applications/$bin
+    fi
 done
